Refresh tokens

+++ authserver/src/internal/dtos/token_response.go
@@ -4,8 +4,8 @@ type TokenResponse struct {
 	AccessToken      string `json:"access_token,omitempty"`
 	IdToken          string `json:"id_token,omitempty"`
 	TokenType        string `json:"token_type,omitempty"`
-	ExpiresIn        int    `json:"expires_in,omitempty"`
+	ExpiresIn        int64  `json:"expires_in,omitempty"`
 	RefreshToken     string `json:"refresh_token,omitempty"`
-	RefreshExpiresIn int    `json:"refresh_expires_in,omitempty"`
+	RefreshExpiresIn int64  `json:"refresh_expires_in,omitempty"`
 	Scope            string `json:"scope,omitempty"`
 }
